requirements.txt

Flask>=2.2
Flask-Cors>=4.0
python-dotenv>=1.0
mysql-connector-python>=8.0
